more test case coverage
