+++ Makefile.defs
@@ -1,6 +1,7 @@
 # including be13_api/Makefile.defs
 BE13_API_SRC= \
         $(BE13_API_DIR)aftimer.h \
+        $(BE13_API_DIR)atomic_histogram.h \
         $(BE13_API_DIR)atomic_set_map.h \
         $(BE13_API_DIR)bulk_extractor_i.h \
         $(BE13_API_DIR)feature_recorder.cpp \
